Update Final Programming Assignment.cpp
Shakey the robot can now pick up items and store them in his inventory. The garden will update in real time to remove the item that shakey is looking at. The robot knows that it cannot pickup the fence wall and other non moveable objects like water, trees and mountains.